+++ partner-chains-cli/src/prepare_configuration/prepare_chain_params.rs
@@ -6,6 +6,7 @@ use crate::config::{config_fields, ConfigFieldDefinition, SidechainParams};
 use crate::io::IOContext;
 use anyhow::{anyhow, Context};
 use serde::de::DeserializeOwned;
+use serde::Deserialize;
 use sidechain_domain::{MainchainAddressHash, UtxoId};
 
 pub fn prepare_chain_params<C: IOContext>(context: &C) -> anyhow::Result<SidechainParams> {
@@ -73,19 +74,16 @@ where
 fn establish_governance_authority(
 	context: &impl IOContext,
 ) -> anyhow::Result<MainchainAddressHash> {
-	let cardano_cli =
-		config_fields::CARDANO_CLI.prompt_with_default_from_file_parse_and_save(context)?;
 	let cardano_payment_verification_key_file =
 		config_fields::CARDANO_PAYMENT_VERIFICATION_KEY_FILE
 			.prompt_with_default_from_file_and_save(context);
 
-	let governance_authority = context.run_command(&format!("{cardano_cli} address key-hash --payment-verification-key-file {cardano_payment_verification_key_file}"))?;
-	let ga = MainchainAddressHash::decode_hex(governance_authority.trim())
-		.map_err(|err| anyhow!(err))
-		.context("Failed to parse governance authority")?;
-	GOVERNANCE_AUTHORITY.save_to_file(&ga, context);
-	context.eprint(&format!("Governance authority has been set to {}", ga));
-	Ok(ga)
+	let governance_authority =
+		get_key_hash_from_file(&cardano_payment_verification_key_file, context)?;
+
+	GOVERNANCE_AUTHORITY.save_to_file(&governance_authority, context);
+	context.eprint(&format!("Governance authority has been set to {}", governance_authority));
+	Ok(governance_authority)
 }
 
 const INTRO: &str = "Now, let's set up the chain parameters: chain id and governance authority. The Pair `(governance authority, chain id)` identifies a partner chain, thus it has to be unique.";
@@ -95,12 +93,52 @@ fn is_gov_auth_valid_prompt() -> String {
 	format!("Is the {} displayed above correct?", GOVERNANCE_AUTHORITY.name)
 }
 
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+struct PaymentKeyFileContent {
+	cbor_hex: String,
+}
+
+const CBOR_KEY_PREFIX: &str = "5820";
+
+fn get_key_hash_from_file(
+	path: &str,
+	context: &impl IOContext,
+) -> anyhow::Result<MainchainAddressHash> {
+	let Some(file_content) = context.read_file(path) else {
+		return Err(anyhow!("Failed to read public key file {path}"));
+	};
+
+	let json = serde_json::from_str::<PaymentKeyFileContent>(&file_content)
+		.map_err(|err| anyhow!("Failed to parse public key file {path}: {err:?}"))?;
+
+	let Some(vkey) = json.cbor_hex.strip_prefix(CBOR_KEY_PREFIX) else {
+		return Err(anyhow!(
+			"Invalid cbor value of public key - missing prefix {CBOR_KEY_PREFIX}: {}",
+			json.cbor_hex
+		));
+	};
+
+	let bytes: [u8; 32] = hex::decode(vkey)
+		.map_err(|err| {
+			anyhow!("Invalid cbor value of public key - not valid hex: {}\n{err:?}", json.cbor_hex)
+		})?
+		.try_into()
+		.map_err(|_| {
+			anyhow!("Invalid cbor value of public key - incorrect length: {}", json.cbor_hex)
+		})?;
+
+	Ok(MainchainAddressHash::from_vkey(bytes))
+}
+
 #[cfg(test)]
 mod tests {
 	use crate::config::config_fields::{
-		CARDANO_CLI, CARDANO_PAYMENT_VERIFICATION_KEY_FILE, CHAIN_ID, GENESIS_COMMITTEE_UTXO,
-		GOVERNANCE_AUTHORITY, THRESHOLD_DENOMINATOR, THRESHOLD_NUMERATOR,
+		CARDANO_PAYMENT_VERIFICATION_KEY_FILE, CHAIN_ID, GENESIS_COMMITTEE_UTXO,
+		GOVERNANCE_AUTHORITY as GOVERNANCE_AUTHORITY_FIELD, THRESHOLD_DENOMINATOR,
+		THRESHOLD_NUMERATOR,
 	};
+	use crate::config::RESOURCES_CONFIG_FILE_PATH;
 	use crate::prepare_configuration::prepare_chain_params::tests::scenarios::silently_fill_legacy_chain_params;
 	use crate::prepare_configuration::prepare_chain_params::{
 		is_gov_auth_valid_prompt, prepare_chain_params, CHAIN_ID_PROMPT,
@@ -114,7 +152,15 @@ mod tests {
 	use sidechain_domain::{MainchainAddressHash, UtxoId};
 	use std::str::FromStr;
 
-	const GOV_AUTH: &str = "0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9";
+	const GOVERNANCE_AUTHORITY: &str = "0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9";
+
+	fn test_vkey_file_json() -> serde_json::Value {
+		serde_json::json!({
+			"type": "PaymentVerificationKeyShelley_ed25519",
+			"description": "Payment Verification Key",
+			"cborHex": "5820516c971f57d5db063161b3240dfa95cdd8030242dd8756c5a003978b4113788c"
+		})
+	}
 
 	mod scenarios {
 		use super::*;
@@ -145,26 +191,23 @@ mod tests {
 	fn happy_path() {
 		let final_chain_config = test_chain_config();
 		let mock_context = MockIOContext::new()
+			.with_file(RESOURCES_CONFIG_FILE_PATH, "{}")
+			.with_json_file("payment.vkey", test_vkey_file_json())
 			.with_expected_io(vec![
-			scenarios::show_intro(),
-			MockIO::prompt(CARDANO_CLI.name, CARDANO_CLI.default, "cardano-cli"),
-			MockIO::file_write_json(
-				CARDANO_CLI.config_file,
-				serde_json::json!({"cardano_cli": "cardano-cli"}),
-			),
-			prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
-			MockIO::run_command("cardano-cli address key-hash --payment-verification-key-file payment.vkey", "76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			save_to_new_file(GOVERNANCE_AUTHORITY, "0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			MockIO::eprint(CHAIN_ID_PROMPT),
-			prompt_and_save_to_existing_file(CHAIN_ID, "0"),
-			silently_fill_legacy_chain_params(final_chain_config),
-            ]);
+				scenarios::show_intro(),
+				prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
+				MockIO::file_read("payment.vkey"),
+				save_to_new_file(GOVERNANCE_AUTHORITY_FIELD, GOVERNANCE_AUTHORITY),
+				MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
+				MockIO::eprint(CHAIN_ID_PROMPT),
+				prompt_and_save_to_existing_file(CHAIN_ID, "0"),
+				silently_fill_legacy_chain_params(final_chain_config),
+			]);
 
 		let result = prepare_chain_params(&mock_context);
 
-		mock_context.no_more_io_expected();
 		assert!(result.is_ok());
+		mock_context.no_more_io_expected();
 	}
 
 	#[test]
@@ -173,29 +216,26 @@ mod tests {
 
 		let initial_chain_config = serde_json::json!({
 			"chain_parameters": {
-				"governance_authority": GOV_AUTH,
+				"governance_authority": GOVERNANCE_AUTHORITY,
 			}
 		});
 
 		let mock_context = MockIOContext::new()
+			.with_file(RESOURCES_CONFIG_FILE_PATH, "{}")
+			.with_json_file("payment.vkey", test_vkey_file_json())
 			.with_json_file(CHAIN_CONFIG_PATH, initial_chain_config).with_expected_io(vec![
-			scenarios::show_intro(),
-			MockIO::file_read(GOVERNANCE_AUTHORITY.config_file),
-			MockIO::eprint(&GOVERNANCE_AUTHORITY.loaded_from_config_msg(&MainchainAddressHash::from_hex_unsafe(GOV_AUTH))),
-			MockIO::prompt_yes_no(&is_gov_auth_valid_prompt(), true, false),
-			MockIO::prompt(CARDANO_CLI.name, CARDANO_CLI.default, "cardano-cli"),
-			MockIO::file_write_json(
-                CARDANO_CLI.config_file,
-                serde_json::json!({"cardano_cli": "cardano-cli"}),
-            ),
-			prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
-			MockIO::run_command("cardano-cli address key-hash --payment-verification-key-file payment.vkey", "76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			save_to_existing_file(GOVERNANCE_AUTHORITY, GOV_AUTH),
-			MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			MockIO::eprint(CHAIN_ID_PROMPT),
-			prompt_and_save_to_existing_file(CHAIN_ID, "0"),
-			silently_fill_legacy_chain_params(final_chain_config),
-        ]);
+				scenarios::show_intro(),
+				MockIO::file_read(GOVERNANCE_AUTHORITY_FIELD.config_file),
+				MockIO::eprint(&GOVERNANCE_AUTHORITY_FIELD.loaded_from_config_msg(&MainchainAddressHash::from_hex_unsafe(GOVERNANCE_AUTHORITY))),
+				MockIO::prompt_yes_no(&is_gov_auth_valid_prompt(), true, false),
+				prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
+				MockIO::file_read("payment.vkey"),
+				save_to_existing_file(GOVERNANCE_AUTHORITY_FIELD, GOVERNANCE_AUTHORITY),
+				MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
+				MockIO::eprint(CHAIN_ID_PROMPT),
+				prompt_and_save_to_existing_file(CHAIN_ID, "0"),
+				silently_fill_legacy_chain_params(final_chain_config),
+			]);
 
 		let result = prepare_chain_params(&mock_context);
 
@@ -207,7 +247,8 @@ mod tests {
 	#[test]
 	fn happy_path_without_overwriting_governance_authority() {
 		let mut final_chain_config = test_chain_config();
-		if let Some(gov_auth) = final_chain_config.pointer_mut(&GOVERNANCE_AUTHORITY.json_pointer())
+		if let Some(gov_auth) =
+			final_chain_config.pointer_mut(&GOVERNANCE_AUTHORITY_FIELD.json_pointer())
 		{
 			*gov_auth = "76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9".into();
 		}
@@ -219,14 +260,15 @@ mod tests {
 		});
 
 		let mock_context = MockIOContext::new()
+			.with_json_file("payment.vkey", test_vkey_file_json())
+			.with_file(RESOURCES_CONFIG_FILE_PATH, "{}")
 			.with_json_file(CHAIN_CONFIG_PATH, initial_chain_config)
 			.with_expected_io(vec![
 				scenarios::show_intro(),
-				MockIO::file_read(GOVERNANCE_AUTHORITY.config_file),
-				MockIO::eprint(
-					&GOVERNANCE_AUTHORITY
-						.loaded_from_config_msg(&MainchainAddressHash::from_hex_unsafe(GOV_AUTH)),
-				),
+				MockIO::file_read(GOVERNANCE_AUTHORITY_FIELD.config_file),
+				MockIO::eprint(&GOVERNANCE_AUTHORITY_FIELD.loaded_from_config_msg(
+					&MainchainAddressHash::from_hex_unsafe(GOVERNANCE_AUTHORITY),
+				)),
 				MockIO::prompt_yes_no(&is_gov_auth_valid_prompt(), true, true),
 				MockIO::eprint(CHAIN_ID_PROMPT),
 				prompt_and_save_to_existing_file(CHAIN_ID, "0"),
@@ -254,22 +296,19 @@ mod tests {
 		});
 
 		let mock_context = MockIOContext::new()
+			.with_json_file("payment.vkey", test_vkey_file_json())
+			.with_file(RESOURCES_CONFIG_FILE_PATH, "{}")
 			.with_json_file(CHAIN_ID.config_file, initial_chain_config).with_expected_io(vec![
-			scenarios::show_intro(),
-			MockIO::file_read(GOVERNANCE_AUTHORITY.config_file),
-			MockIO::prompt(CARDANO_CLI.name, CARDANO_CLI.default, "cardano-cli"),
-			MockIO::file_write_json(
-				CARDANO_CLI.config_file,
-				serde_json::json!({"cardano_cli": "cardano-cli"}),
-			),
-			prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
-			MockIO::run_command("cardano-cli address key-hash --payment-verification-key-file payment.vkey", "76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			save_to_existing_file(GOVERNANCE_AUTHORITY,  "0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			MockIO::eprint(CHAIN_ID_PROMPT),
-			prompt_with_default_and_save_to_existing_file(CHAIN_ID, Some("1"),"2"),
-			silently_fill_legacy_chain_params(final_chain_config),
-		]);
+				scenarios::show_intro(),
+				MockIO::file_read(GOVERNANCE_AUTHORITY_FIELD.config_file),
+				prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
+				MockIO::file_read("payment.vkey"),
+				save_to_existing_file(GOVERNANCE_AUTHORITY_FIELD, GOVERNANCE_AUTHORITY),
+				MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
+				MockIO::eprint(CHAIN_ID_PROMPT),
+				prompt_with_default_and_save_to_existing_file(CHAIN_ID, Some("1"),"2"),
+				silently_fill_legacy_chain_params(final_chain_config),
+			]);
 
 		let result = prepare_chain_params(&mock_context);
 
@@ -303,31 +342,28 @@ mod tests {
 		});
 
 		let mock_context = MockIOContext::new()
+			.with_file(RESOURCES_CONFIG_FILE_PATH, "{}")
 			.with_json_file(GENESIS_COMMITTEE_UTXO.config_file ,initial_chain_config)
+			.with_json_file("payment.vkey", test_vkey_file_json())
 			.with_expected_io(vec![
-			scenarios::show_intro(),
-			MockIO::file_read(GOVERNANCE_AUTHORITY.config_file),
-			MockIO::prompt(CARDANO_CLI.name, CARDANO_CLI.default, "cardano-cli"),
-			MockIO::file_write_json(
-				CARDANO_CLI.config_file,
-				serde_json::json!({"cardano_cli": "cardano-cli"}),
-			),
-			prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
-			MockIO::run_command("cardano-cli address key-hash --payment-verification-key-file payment.vkey", "76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			save_to_existing_file(GOVERNANCE_AUTHORITY, "0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
-			MockIO::eprint(CHAIN_ID_PROMPT),
-			MockIO::file_read(CHAIN_ID.config_file),
-			MockIO::prompt(CHAIN_ID.name, CHAIN_ID.default, "0"),
-			MockIO::file_read(CHAIN_ID.config_file),
-			MockIO::file_write_json(CHAIN_ID.config_file, final_chain_config),
-			MockIO::file_read(GENESIS_COMMITTEE_UTXO.config_file),
-			MockIO::eprint(&GENESIS_COMMITTEE_UTXO.loaded_from_config_msg(&UtxoId::from_str(genesis_committee_utxo).unwrap())),
-			MockIO::file_read(THRESHOLD_NUMERATOR.config_file),
-			MockIO::eprint(&THRESHOLD_NUMERATOR.loaded_from_config_msg(&threshold_numerator)),
-			MockIO::file_read(THRESHOLD_DENOMINATOR.config_file),
-			MockIO::eprint(&THRESHOLD_DENOMINATOR.loaded_from_config_msg(&threshold_denominator)),
-		]);
+				scenarios::show_intro(),
+				MockIO::file_read(GOVERNANCE_AUTHORITY_FIELD.config_file),
+				prompt_and_save_to_existing_file(CARDANO_PAYMENT_VERIFICATION_KEY_FILE, "payment.vkey"),
+				MockIO::file_read("payment.vkey"),
+				save_to_existing_file(GOVERNANCE_AUTHORITY_FIELD, GOVERNANCE_AUTHORITY),
+				MockIO::eprint("Governance authority has been set to 0x76da17b2e3371ab7ca88ce0500441149f03cc5091009f99c99c080d9"),
+				MockIO::eprint(CHAIN_ID_PROMPT),
+				MockIO::file_read(CHAIN_ID.config_file),
+				MockIO::prompt(CHAIN_ID.name, CHAIN_ID.default, "0"),
+				MockIO::file_read(CHAIN_ID.config_file),
+				MockIO::file_write_json(CHAIN_ID.config_file, final_chain_config),
+				MockIO::file_read(GENESIS_COMMITTEE_UTXO.config_file),
+				MockIO::eprint(&GENESIS_COMMITTEE_UTXO.loaded_from_config_msg(&UtxoId::from_str(genesis_committee_utxo).unwrap())),
+				MockIO::file_read(THRESHOLD_NUMERATOR.config_file),
+				MockIO::eprint(&THRESHOLD_NUMERATOR.loaded_from_config_msg(&threshold_numerator)),
+				MockIO::file_read(THRESHOLD_DENOMINATOR.config_file),
+				MockIO::eprint(&THRESHOLD_DENOMINATOR.loaded_from_config_msg(&threshold_denominator)),
+			]);
 
 		let result = prepare_chain_params(&mock_context);